components/client/requirements.txt
pandas
requests
opencv-python==4.2.0.34
numpy
v3io
v3io_frames
pyyaml